+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Changes in the next release_
 
+### Fixed
+- Automatically migrate old configuration file at startup and configuration flow, otherwise the device must be paired again.
+
 ---
 
 ## v0.13.1 - 2024-03-07